+++ app/models/book.rb
@@ -1,3 +1,4 @@
 class Book < ApplicationRecord
-  has_many :ratings
+  has_many :ratings, :dependent => :destroy
+  
 end
