+++ app/models/user.rb
@@ -1,5 +1,5 @@
 class User < ApplicationRecord
-  has_many :ratings
+  has_many :ratings, :dependent => :destroy
   has_secure_password
   validates :password, length: { minimum: 6 }
   validates :username, presence: true, uniqueness: true
