+++ client/src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import Layout from "./layouts/Layout"
 import { Switch, Route, useHistory } from "react-router-dom"
-import Login from "./screens/Login"
+import Login from "./screens/Login/Login"
 import Register from "./screens/Register"
 import {useState, useEffect} from "react"
 import { loginUser, registerUser, removeToken, verifyUser } from './services/auth';
